split renderer and register logic
